Make generic Scraper class
