+++ scraper/boomkat.py
@@ -4,12 +4,10 @@ from dataclasses import dataclass
 from typing import Union
 
 import bs4
-import requests
+from base_scraper import Scraper
 from bs4 import BeautifulSoup
 
 logger = logging.getLogger(__name__)
-
-_CACHE_DIR = "../.cache"
 
 
 @dataclass
@@ -27,7 +25,7 @@ class BoomkatChart:
     items: list[BoomkatRelease]
 
 
-class BoomkatScraper:
+class BoomkatScraper(Scraper):
     _BASE_URL: str = "https://boomkat.com"
 
     # Matches a title and label in the format "TITLE (LABEL)"
@@ -35,36 +33,6 @@ class BoomkatScraper:
 
     def __init__(self):
         pass
-
-    def _get_cache_path(self, url: str) -> str:
-        return f"{_CACHE_DIR}/{url.replace('/', '_')}"
-
-    def _get_cached(self, url: str) -> str:
-        cache_path = self._get_cache_path(url)
-        try:
-            with open(cache_path, "r") as f:
-                return f.read()
-        except FileNotFoundError:
-            return None
-
-    def _cache(self, url: str, content: str):
-        cache_path = self._get_cache_path(url)
-        with open(cache_path, "w") as f:
-            f.write(content)
-
-    # write a series of functions to retrieve a url from the cache, or if it doesn't exist, fetch it from the internet
-    def _get_with_cache(self, url: str) -> str:
-        cached = self._get_cached(url)
-        if cached:
-            logger.debug(f"Using cached content for {url}")
-            return cached
-        response = requests.get(url, verify=False)
-        if response.status_code != 200:
-            raise Exception(f"Failed to fetch {url}")
-        content = response.text
-        logger.debug(f"Caching content for {url}")
-        self._cache(url, content)
-        return content
 
     def _get_charts_from_root(self) -> bs4.element.ResultSet:
         # TODO make root dynamic (i.e. query previous years)
