+++ scraper/scraper.py
@@ -1,84 +1,42 @@
 import logging
-from dataclasses import dataclass
 
-import urllib3
-from boomkat import BoomkatChart, BoomkatScraper
-
-logging.basicConfig(
-    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
-)
+import requests
 
 logger = logging.getLogger(__name__)
 
-# TODO understand why it doesn't trust boomkat certificate
-urllib3.disable_warnings(category=urllib3.exceptions.InsecureRequestWarning)
+
+_CACHE_DIR = ".cache"
 
 
-@dataclass
-class Title:
-    name: str
-    charts: list[str]
+class Scraper:
+    def __init__(self, url: str):
+        self.url = url
 
+    def _get_cache_path(self, url: str) -> str:
+        return f"{_CACHE_DIR}/{url.replace('/', '_')}"
 
-@dataclass
-class Artist:
-    name: str
-    titles: dict[str, Title]
+    def _get_cached(self, url: str) -> str:
+        cache_path = self._get_cache_path(url)
+        try:
+            with open(cache_path, "r") as f:
+                return f.read()
+        except FileNotFoundError:
+            return None
 
+    def _cache(self, url: str, content: str):
+        cache_path = self._get_cache_path(url)
+        with open(cache_path, "w") as f:
+            f.write(content)
 
-@dataclass
-class Release:
-    artist: str
-    title: str
-    charts: list[str]
-
-    def appearances(self) -> int:
-        return len(self.charts)
-
-
-def _group_by_artist(charts: list[BoomkatChart]) -> dict[str, Artist]:
-    # artists = defaultdict(Artist)
-    artists = {}
-    for chart in charts:
-        for item in chart.items:
-            artist_lower = item.artist.lower()
-            title_lower = item.title.lower()
-            if artist_lower not in artists:
-                artists[artist_lower] = Artist(
-                    name=item.artist,
-                    titles={},
-                )
-            if title_lower not in artists[artist_lower].titles:
-                artists[artist_lower].titles[title_lower] = Title(
-                    name=item.title,
-                    charts=[],
-                )
-            artists[artist_lower].titles[title_lower].charts.append(chart.name)
-    return artists
-
-
-def _flatten_to_releases(artists: dict[str, Artist]) -> list[Release]:
-    releases = []
-    for artist_lower, artist in artists.items():
-        for title_lower, title in artist.titles.items():
-            releases.append(
-                Release(
-                    artist=artist.name,
-                    title=title.name,
-                    charts=title.charts,
-                )
-            )
-    return releases
-
-
-if __name__ == "__main__":
-    scraper = BoomkatScraper()
-    charts = scraper.scrape()
-    artists = _group_by_artist(charts)
-    releases = _flatten_to_releases(artists)
-
-    popular_releases = sorted(releases, key=lambda x: x.appearances(), reverse=True)
-    for release in popular_releases[0:10]:
-        logger.info(
-            f'"{release.artist} - {release.title}" appears in {release.appearances()} charts'
-        )
+    def _get_with_cache(self, url: str) -> str:
+        cached = self._get_cached(url)
+        if cached:
+            logger.debug(f"Using cached content for {url}")
+            return cached
+        response = requests.get(url, verify=False)
+        if response.status_code != 200:
+            raise Exception(f"Failed to fetch {url}")
+        content = response.text
+        logger.debug(f"Caching content for {url}")
+        self._cache(url, content)
+        return content
